First completed todolist. [First PHP/MySQL/Tailwind Application (01/24/2021)]

+++ includes/function/delete.php
@@ -1,10 +1,15 @@
 <?php
 require 'db.php';
 
-$id = $_GET['id'];
-$query = 'DELETE FROM todos WHERE id=:id';
-$statement = $conn->prepare($query);
-
-if($statement->execute([':id' => $id])){
-    header("Location: /");
+if(isset($_GET['id'])){
+    $id = $_GET['id'];
+    $query = 'DELETE FROM todos WHERE id=:id';
+    $statement = $conn->prepare($query);
+    
+    if($statement->execute([':id' => $id])){
+        header("Location: /Todolist/index.php");
+    }
+}else{
+    echo 'error';
+    header("Location: /Todolist/index.php");
 }
